+++ src/EventListener/Doctrine/DoctrineProviderListener.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace AutoMapper\EventListener\Doctrine;
 
 use AutoMapper\Event\GenerateMapperEvent;
+use AutoMapper\Metadata\Provider;
 use AutoMapper\Provider\Doctrine\DoctrineProvider;
 use Doctrine\Persistence\ObjectManager;
 
@@ -21,6 +22,6 @@ final readonly class DoctrineProviderListener
             return;
         }
 
-        $event->provider ??= DoctrineProvider::class;
+        $event->provider ??= new Provider(Provider::TYPE_SERVICE, DoctrineProvider::class);
     }
 }