+++ src/Generator/PropertyConditionsGenerator.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace AutoMapper\Generator;
 
+use AutoMapper\AttributeReference\AttributeInstance;
+use AutoMapper\AttributeReference\Reference;
 use AutoMapper\Exception\CompileException;
 use AutoMapper\MapperContext;
 use AutoMapper\Metadata\GeneratorMetadata;
@@ -16,6 +18,7 @@ use PhpParser\Node\Stmt;
 use PhpParser\Parser;
 use PhpParser\ParserFactory;
 use Symfony\Component\ExpressionLanguage\ExpressionLanguage;
+use Symfony\Component\ObjectMapper\ConditionCallableInterface;
 
 use function AutoMapper\PhpParser\create_expr_array_item;
 use function AutoMapper\PhpParser\create_scalar_int;
@@ -246,6 +249,32 @@ final readonly class PropertyConditionsGenerator
         }
 
         $callableName = null;
+        $value = $metadata->variableRegistry->getSourceInput();
+        $input = null;
+
+        // use read accessor
+        if ($propertyMetadata->source->accessor !== null) {
+            $input = $propertyMetadata->source->accessor->getExpression($metadata->variableRegistry->getSourceInput());
+        }
+
+        if ($propertyMetadata->if instanceof Reference) {
+            $refExpr = $propertyMetadata->if->getReferenceExpression();
+
+            /** (AttributeInstance::get($attributeClassName, $index, $reference)->transformer)(...) */
+            return new Expr\FuncCall(new Expr\PropertyFetch(new Expr\StaticCall(
+                new Name\FullyQualified(AttributeInstance::class),
+                'get',
+                [
+                    new Arg(new Scalar\String_($propertyMetadata->if->attributeClassName)),
+                    new Arg($refExpr),
+                    new Arg(new Scalar\Int_($propertyMetadata->if->attributeIndex)),
+                ]
+            ), 'if'), [
+                new Arg($value),
+                new Arg($input ?? new Expr\ConstFetch(new Name('null'))),
+                new Arg(new Expr\Variable('context')),
+            ]);
+        }
 
         if (\is_callable($propertyMetadata->if, false, $callableName)) {
             if (\function_exists($callableName)) {
@@ -257,12 +286,22 @@ final readonly class PropertyConditionsGenerator
                     return new Expr\FuncCall(
                         new Name($callableName),
                         [
-                            new Arg(new Expr\Variable('value')),
+                            new Arg($value),
                         ]
                     );
                 }
 
-                if ($argumentsCount > 2) {
+                if ($argumentsCount === 2) {
+                    return new Expr\FuncCall(
+                        new Name($callableName),
+                        [
+                            new Arg($value),
+                            new Arg(new Expr\Variable('context')),
+                        ]
+                    );
+                }
+
+                if ($argumentsCount > 3) {
                     throw new CompileException('Callable condition must have 1 or 2 arguments required, but it has ' . $argumentsCount);
                 }
             }
@@ -270,7 +309,8 @@ final readonly class PropertyConditionsGenerator
             return new Expr\FuncCall(
                 new Name($callableName),
                 [
-                    new Arg(new Expr\Variable('value')),
+                    new Arg($value),
+                    new Arg($input ?? new Expr\ConstFetch(new Name('null'))),
                     new Arg(new Expr\Variable('context')),
                 ]
             );
@@ -284,18 +324,41 @@ final readonly class PropertyConditionsGenerator
                     new Name\FullyQualified($metadata->mapperMetadata->source),
                     $propertyMetadata->if,
                     [
-                        new Arg(new Expr\Variable('value')),
+                        new Arg($value),
+                        new Arg($input ?? new Expr\ConstFetch(new Name('null'))),
                         new Arg(new Expr\Variable('context')),
                     ]
                 );
             }
 
             return new Expr\MethodCall(
-                new Expr\Variable('value'),
+                $metadata->variableRegistry->getSourceInput(),
                 $propertyMetadata->if,
                 [
-                    new Arg(new Expr\Variable('value')),
+                    // pass null as value if there is no read accessor
+                    new Arg($input ?? new Expr\ConstFetch(new Name('null'))),
                     new Arg(new Expr\Variable('context')),
+                ]
+            );
+        }
+
+        if (class_exists($propertyMetadata->if) && is_subclass_of($propertyMetadata->if, ConditionCallableInterface::class)) {
+            return new Expr\MethodCall(
+                new Expr\NullsafeMethodCall(
+                    new Expr\PropertyFetch(
+                        new Expr\Variable('this'),
+                        'serviceLocator'
+                    ),
+                    'get',
+                    [
+                        new Arg(new Scalar\String_($propertyMetadata->if)),
+                    ]
+                ),
+                '__invoke',
+                [
+                    new Arg($input ?? $value),
+                    new Arg($value),
+                    new Arg(new Expr\Variable('result')),
                 ]
             );
         }