+++ src/Transformer/ReferenceTransformer.php
@@ -6,7 +6,6 @@ namespace AutoMapper\Transformer;
 
 use AutoMapper\AttributeReference\AttributeInstance;
 use AutoMapper\AttributeReference\Reference;
-use AutoMapper\AttributeReference\ReflectionReference;
 use AutoMapper\Generator\UniqueVariableScope;
 use AutoMapper\Metadata\PropertyMetadata;
 use PhpParser\Node\Arg;
@@ -18,40 +17,26 @@ final readonly class ReferenceTransformer implements TransformerInterface
 {
     public function __construct(
         private Reference $reference,
+        private bool $objectMapperTransformer = false,
     ) {
     }
 
     public function transform(Expr $input, Expr $target, PropertyMetadata $propertyMapping, UniqueVariableScope $uniqueVariableScope, Expr $source, ?Expr $existingValue = null): array
     {
-        if ($this->reference->methodName) {
-            /** ReflectionReference::fromMethod($className, $methodName) */
-            $reflectionReferenceExpr = new Expr\StaticCall(
-                new Name\FullyQualified(ReflectionReference::class),
-                'fromMethod',
-                [
-                    new Arg(new Scalar\String_($this->reference->className)),
-                    new Arg(new Scalar\String_($this->reference->methodName)),
-                ]
-            );
-        } elseif ($this->reference->propertyName) {
-            /** ReflectionReference::fromProperty($className, $propertyName) */
-            $reflectionReferenceExpr = new Expr\StaticCall(
-                new Name\FullyQualified(ReflectionReference::class),
-                'fromProperty',
-                [
-                    new Arg(new Scalar\String_($this->reference->className)),
-                    new Arg(new Scalar\String_($this->reference->propertyName)),
-                ]
-            );
+        $reflectionReferenceExpr = $this->reference->getReferenceExpression();
+
+        if ($this->objectMapperTransformer) {
+            $args = [
+                new Arg($input),
+                new Arg($source),
+                new Arg(new Expr\ConstFetch(new Name('null'))),
+            ];
         } else {
-            /** ReflectionReference::fromClass($className) */
-            $reflectionReferenceExpr = new Expr\StaticCall(
-                new Name\FullyQualified(ReflectionReference::class),
-                'fromClass',
-                [
-                    new Arg(new Scalar\String_($this->reference->className)),
-                ]
-            );
+            $args = [
+                new Arg($input),
+                new Arg($source),
+                new Arg(new Expr\Variable('context')),
+            ];
         }
 
         /** (AttributeInstance::get($attributeClassName, $index, $reference)->transformer)(...) */
@@ -64,11 +49,7 @@ final readonly class ReferenceTransformer implements TransformerInterface
                     new Arg($reflectionReferenceExpr),
                     new Arg(new Scalar\Int_($this->reference->attributeIndex)),
                 ]
-            ), 'transformer'), [
-                new Arg($input),
-                new Arg($source),
-                new Arg(new Expr\Variable('context')),
-            ]), [],
+            ), $this->objectMapperTransformer ? 'transform' : 'transformer'), $args), [],
         ];
     }
 }