+++ tests/ObjectMapper/Fixtures/EmbeddedMapping/Address.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace AutoMapper\Tests\ObjectMapper\Fixtures\EmbeddedMapping;
+
+class Address
+{
+    public string $zipcode;
+    public string $city;
+}