+++ tests/ObjectMapper/Fixtures/ReadOnlyPromotedProperty/ReadOnlyPromotedPropertyAMapped.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace AutoMapper\Tests\ObjectMapper\Fixtures\ReadOnlyPromotedProperty;
+
+final class ReadOnlyPromotedPropertyAMapped
+{
+    public function __construct(
+        public ReadOnlyPromotedPropertyBMapped $b,
+        public string $var1,
+    ) {
+    }
+}