+++ tests/ObjectMapper/Fixtures/TransformCollection/TransformCollectionB.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace AutoMapper\Tests\ObjectMapper\Fixtures\TransformCollection;
+
+class TransformCollectionB
+{
+    /** @var TransformCollectionD[] */
+    public array $foo;
+}